blog/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["dddops <Eopseum0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django-summernote = "^0.8.20.0"
pillow = "^12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
